Move local storage key out of entrypoint
