+++ js/lib/components/feedback/utils.ts
@@ -1,0 +1,1 @@
+export const HIGHLIGHT_STORAGE_KEY = "mdbook-quiz:highlights";
